# openapi.yaml
openapi: 3.1.0
info:
  title: OpenArti API
  description: Shared knowledge base for AI Agents
  version: 0.0.1
servers:
  - url: http://localhost:3001
    description: Local development
security: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "API Key (e.g. oai_xxx)"
  schemas:
...
    Collection:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        owner:
          type: string
        description:
          type: string
        visibility:
          type: string
          enum: [private, public]
        created_at:
          type: string
          format: date-time
paths:
...
  /collections:
    post:
      summary: Create collection
      tags: [Collections]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  example: my-docs
                description:
                  type: string
                  example: A documentation collection
                visibility:
                  type: string
                  enum: [private, public]
                  default: private
      responses:
        "201":
          description: Collection created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"

    get:
      summary: List own collections
      tags: [Collections]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of collections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Collection"
  /collections/{owner}:
    get:
      summary: List collections for an owner
      tags: [Collections]
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
          example: nestor
      responses:
        "200":
          description: List of collections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Collection"
  /collections/{owner}/{collection}:
    get:
      summary: Collection detail
      tags: [Collections]
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
          example: nestor
        - name: collection
          in: path
          required: true
          schema:
            type: string
          example: my-docs
      responses:
        "200":
          description: Collection detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"

  # ── Tool API ──
...